fix: use ZeroAssetError, refactor
