+++ pkg/types/indicator.go
@@ -998,51 +998,6 @@ func Rolling(a Series, window int) *RollingResult {
 	return &RollingResult{a, window}
 }
 
-type FilterResult struct {
-	a      Series
-	b      func(int, float64) bool
-	length int
-	c      []int
-}
-
-func (f *FilterResult) Last() float64 {
-	return f.Index(0)
-}
-
-func (f *FilterResult) Index(j int) float64 {
-	if j >= f.length {
-		return 0
-	}
-	if len(f.c) > j {
-		return f.a.Index(f.c[j])
-	}
-	l := f.a.Length()
-	k := len(f.c)
-	i := 0
-	if k > 0 {
-		i = f.c[k-1] + 1
-	}
-	for ; i < l; i++ {
-		tmp := f.a.Index(i)
-		if f.b(i, tmp) {
-			f.c = append(f.c, i)
-			if j == k {
-				return tmp
-			}
-			k++
-		}
-	}
-	return 0
-}
-
-func (f *FilterResult) Length() int {
-	return f.length
-}
-
-func Filter(a Series, b func(i int, value float64) bool, length int) SeriesExtend {
-	return NewSeries(&FilterResult{a, b, length, nil})
-}
-
 type SigmoidResult struct {
 	a Series
 }
